1.fix all bugs 2.add features 3.refactoring
Took 26 hours 15 minutes

+++ models/order.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Boolean, String
+from sqlalchemy import Column, Integer, Boolean, String, BigInteger
 from database import Base
 
 
@@ -9,7 +9,7 @@ class Order(Base):
     __tablename__ = 'order'
     id = Column(Integer, primary_key=True, index=True)
     amount = Column(Integer, default=0)
-    user_telegram = Column(Integer, nullable=False)
+    user_telegram = Column(BigInteger, nullable=False)
     user_name = Column(String, nullable=False)
     post_id = Column(Integer, nullable=False)
     paid = Column(Boolean, default=False)
